Make Configurable sealed
