+++ src/config/redirect.rs
@@ -1,0 +1,52 @@
+use super::SetOpt;
+use curl::easy::Easy2;
+
+/// Describes a policy for handling server redirects.
+///
+/// The default is to not follow redirects.
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub enum RedirectPolicy {
+    /// Do not apply any special treatment to redirect responses. The response
+    /// will be returned as-is and redirects will not be followed.
+    ///
+    /// This is the default policy.
+    None,
+    /// Follow all redirects automatically.
+    Follow,
+    /// Follow redirects automatically up to a maximum number of redirects.
+    Limit(u32),
+}
+
+impl Default for RedirectPolicy {
+    fn default() -> Self {
+        RedirectPolicy::None
+    }
+}
+
+impl SetOpt for RedirectPolicy {
+    fn set_opt<H>(&self, easy: &mut Easy2<H>) -> Result<(), curl::Error> {
+        match self {
+            RedirectPolicy::Follow => {
+                easy.follow_location(true)?;
+            }
+            RedirectPolicy::Limit(max) => {
+                easy.follow_location(true)?;
+                easy.max_redirections(*max)?;
+            }
+            RedirectPolicy::None => {
+                easy.follow_location(false)?;
+            }
+        }
+
+        Ok(())
+    }
+}
+
+#[derive(Clone, Copy, Debug)]
+pub(crate) struct AutoReferer;
+
+impl SetOpt for AutoReferer {
+    fn set_opt<H>(&self, easy: &mut Easy2<H>) -> Result<(), curl::Error> {
+        easy.autoreferer(true)
+    }
+}
